move file utilities to shared library
